Fix CI build failure in test_socket_factory.cpp
- Add missing #include <atomic> header for std::atomic<uint16_t>
- Fix Port constructor usage with explicit cast to uint16_t
- Port constructor is explicit, requires static_cast<uint16_t>()
- Resolves MinGW compilation errors on GitHub Actions
- All tests now pass on Linux, macOS, and Windows CI

+++ tests/test_socket_factory.cpp
@@ -13,6 +13,7 @@
 #include <cstring>
 #include <thread>
 #include <chrono>
+#include <atomic>
 
 using namespace aiSocks;
 
@@ -106,7 +107,7 @@ int main() {
         REQUIRE(srvPort.load() != 0);
 
         auto clt_result = SocketFactory::createTcpClient(AddressFamily::IPv4,
-            ConnectArgs{"127.0.0.1", Port{srvPort.load()}, Milliseconds{2000}});
+            ConnectArgs{"127.0.0.1", Port{static_cast<uint16_t>(srvPort.load())}, Milliseconds{2000}});
         REQUIRE(clt_result.isSuccess());
         auto& client = clt_result.value();
         REQUIRE(client.isValid());
